minor bugs in parser fixed

+++ CompleteCylinder.js
@@ -7,11 +7,11 @@ var degToRad = Math.PI / 180.0;
 
 function CompleteCylinder(scene, coords){
    CGFobject.call(this,scene);
-   
-   this.height = parseInt(coords[0]);
+
+   this.height = parseFloat(coords[0]);
    this.slices = parseInt(coords[4]);
-   this.topRadius = parseInt(coords[2]);
-   this.bottomRadius = parseInt(coords[1]);
+   this.topRadius = parseFloat(coords[2]);
+   this.bottomRadius = parseFloat(coords[1]);
    this.topCap = parseInt(coords[5]);
    this.bottomCap = parseInt(coords[6]);
 
@@ -26,19 +26,19 @@ CompleteCylinder.prototype.constructor=CompleteCylinder;
 
 
 CompleteCylinder.prototype.display = function(){
-   
+
     this.scene.pushMatrix();
     this.cylinder.display();
     this.scene.popMatrix();
-      
+
    if(this.topCap==1){
-       
+
     this.scene.pushMatrix()
     this.scene.translate(0,0,this.height);
     this.circleTop.display();
     this.scene.popMatrix();
     }
-   
+
    if(this.bottomCap==1){
 
     this.scene.pushMatrix();
@@ -52,5 +52,5 @@ CompleteCylinder.prototype.display = function(){
 CompleteCylinder.prototype.setTextCoords = function(s,t){
     this.circleTop.setTextCoords(s, t);
     this.circleBottom.setTextCoords(s, t);
-    this.cylinder.setTextCoords(s, t); 
+    this.cylinder.setTextCoords(s, t);
 };
